% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exif.R
\name{call_exiftool}
\alias{call_exiftool}
\alias{exiftool_version}
\title{Call exiftool from R}
\usage{
call_exiftool(args = NULL, fnames = NULL, intern = FALSE, ...)

exiftool_version()
}
\arguments{
\item{args}{List of non-shell quoted arguments (e.g. \code{-n
-csv})}

\item{fnames}{Character vector giving one or more file paths}

\item{intern}{\code{TRUE} if output should be returned as a
character vector.}

\item{...}{Additional arguments to be passed to \code{system()}}
}
\value{
The exit code if \code{intern=FALSE}, or the standard
    output as a character vector if \code{intern=TRUE}.
}
\description{
Uses \code{system()} to run a basic call to \code{exiftool}.
}
\examples{
call_exiftool()
exiftool_version()

}
