% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exif_read.R
\name{exif_call}
\alias{exif_call}
\alias{exif_version}
\title{Call ExifTool from R}
\usage{
exif_call(args = NULL, fnames = NULL, intern = FALSE, ...)

exif_version()
}
\arguments{
\item{args}{List of non-shell quoted arguments (e.g. \code{-n
-csv})}

\item{fnames}{A character vector giving one or more file paths.}

\item{intern}{\code{TRUE} if output should be returned as a
character vector. Default value is \code{FALSE}.}

\item{...}{Additional arguments to be passed to \code{system()}.}
}
\value{
The exit code if \code{intern=FALSE}, or the standard
    output as a character vector if \code{intern=TRUE}.
}
\description{
Uses \code{system()} to run a basic call to \code{exiftool}.
}
\examples{
\dontrun{
exif_call()
exif_version()
}
}
