% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_exiftool}
\alias{install_exiftool}
\title{Download and install the current version of ExifTool}
\usage{
install_exiftool(install_location = NULL, win_exe = NULL,
  quiet = FALSE)
}
\arguments{
\item{install_location}{Path to the directory into which ExifTool
should be installed. If \code{NULL} (the default),
installation will be into the (initially empty)
\code{exiftool} folder in the \pkg{exiftoolr} package's
directory tree.}

\item{win_exe}{Logical, only used on Windows machines. Should we
install the standalone ExifTool Windows executable or the
ExifTool Perl library?  (The latter relies, for its execution,
on an existing installation of Perl being present on the
user's machine.)  If set to \code{NULL} (the default), the
function installs the Windows executable on Windows machines
and the Perl library on other operating systems.}

\item{quiet}{Logical.  Should function should be chatty?}
}
\description{
Download and install the current version of ExifTool
}
