% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_exiftoolr}
\alias{configure_exiftoolr}
\alias{configure_perl}
\title{Configure Perl, ExifTool}
\usage{
configure_exiftoolr(command = NULL, perl_path = NULL,
  allow_win_exe = TRUE, quiet = FALSE)

configure_perl(perl_path = NULL, quiet = FALSE)
}
\arguments{
\item{command}{Character string giving the exiftool command.}

\item{perl_path}{Path to a Perl executable.}

\item{quiet}{Should function should be chatty?}
}
\value{
The exiftool command, invisibly
}
\description{
Configure Perl, ExifTool
}
