% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_exiftoolr}
\alias{configure_exiftoolr}
\alias{configure_perl}
\title{Configure perl, ExifTool}
\usage{
configure_exiftoolr(command = NULL, perl_path = NULL,
  allow_win_exe = TRUE, quiet = FALSE)

configure_perl(perl_path = NULL, quiet = FALSE)
}
\arguments{
\item{command}{The exiftool command or location of exiftool.pl}

\item{perl_path}{The path to the perl executable}

\item{quiet}{Use quiet = FALSE to display status updates}

\item{install_url}{The url from which exiftool could be installed}

\item{install_location}{The location to install exiftool}
}
\value{
The exiftool command, invisibly
}
\description{
Configure perl, ExifTool
}
