% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_exiftoolr}
\alias{configure_exiftoolr}
\title{Configure package to point to ExifTool and/or Perl}
\usage{
configure_exiftoolr(
  command = NULL,
  perl_path = NULL,
  allow_win_exe = TRUE,
  quiet = FALSE
)
}
\arguments{
\item{command}{Character string giving the exiftool command.}

\item{perl_path}{Path to a Perl executable.}

\item{allow_win_exe}{Logical. If running on a Windows machine, and if a
standalone exiftool executable is available, should it be used?}

\item{quiet}{Logical. Should function should be chatty?}
}
\value{
A character string giving the exiftool command, returned invisibly.
}
\description{
Configure package to point to ExifTool and/or Perl
}
