% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_exiftool}
\alias{install_exiftool}
\title{Install ExifTool on the local machine}
\usage{
install_exiftool(install_location = NULL, win_exe = NULL, quiet = FALSE)
}
\arguments{
\item{install_location}{Path to directory into which exiftool
should be installed. If \code{NULL} (the default),
installation will be into the \pkg{exiftoolr} package's
directory tree.}

\item{win_exe}{Logical. On Windows machines, should we install a
standalone Windows executable or the exiftool Perl library?
(The latter depends on an external installation of Perl.)  If
\code{NULL} (the default) the function installs Windows
executable on Windows machines and Perl library on other
operating systems.}

\item{quiet}{Logical.}
}
\description{
Install ExifTool on the local machine
}
